Add game logic

+++ server/src/main/java/entity/Match.java
@@ -14,22 +14,22 @@ import lombok.ToString;
 @ToString
 public class Match {
     private final long id;
-    private final long user1Id;
-    private final long user2Id;
+    private final int user1Id;
+    private final int user2Id;
     private int user1Score = 0;
     private int user2Score = 0;
-    private Long winnerId = null;
+    private Integer winnerId = null;
     private MatchStatus status = MatchStatus.IN_PROGRESS;
     private LocalDateTime startedAt = LocalDateTime.now();
     private LocalDateTime finishedAt;
 
     public Match(
             long id,
-            long user1Id,
-            long user2Id,
+            int user1Id,
+            int user2Id,
             int user1Score,
             int user2Score,
-            Long winnerId,
+            Integer winnerId,
             MatchStatus status
     ) {
         this.id = id;
