+++ shared/src/main/java/dto/request/MatchFilter.java
@@ -3,8 +3,8 @@ package dto.request;
 import java.time.LocalDateTime;
 
 public record MatchFilter(
-        Long userId1,
-        Long userId2,
+        Integer userId1,
+        Integer userId2,
         String status,
         LocalDateTime startedAtFrom,
         LocalDateTime startedAtTo,
